# chore(lint): make ESLint concurrency configurable via pnpm config