--- a/.env_example
+++ b/.env_example
@@ -33,3 +33,6 @@ DISABLE_VUE_PLUGINS=false
 # Algolia credentials required for developing with the new custom node manager.
 ALGOLIA_APP_ID=4E0RO38HS8
 ALGOLIA_API_KEY=684d998c36b67a9a9fce8fc2d8860579
+
+# Optional: control ESLint concurrency used by pnpm scripts.
+# ESLINT_CONCURRENCY=4