--- a/.env_example
+++ b/.env_example
@@ -34,5 +34,3 @@ DISABLE_VUE_PLUGINS=false
 ALGOLIA_APP_ID=4E0RO38HS8
 ALGOLIA_API_KEY=684d998c36b67a9a9fce8fc2d8860579
 
-# Optional: control ESLint concurrency used by pnpm scripts.
-# ESLINT_CONCURRENCY=auto