--- a/package.json
+++ b/package.json
@@ -25,10 +25,10 @@
     "preinstall": "npx only-allow pnpm",
     "prepare": "husky || true && git config blame.ignoreRevsFile .git-blame-ignore-revs || true",
     "preview": "nx preview",
-    "lint": "tsx scripts/run-eslint.ts --cache",
-    "lint:fix": "tsx scripts/run-eslint.ts --fix --cache",
-    "lint:no-cache": "tsx scripts/run-eslint.ts",
-    "lint:fix:no-cache": "tsx scripts/run-eslint.ts --fix",
+    "lint": "eslint src --cache --concurrency=$npm_package_config_eslint_concurrency",
+    "lint:fix": "eslint src --fix --cache --concurrency=$npm_package_config_eslint_concurrency",
+    "lint:no-cache": "eslint src --concurrency=$npm_package_config_eslint_concurrency",
+    "lint:fix:no-cache": "eslint src --fix --concurrency=$npm_package_config_eslint_concurrency",
     "knip": "knip --cache",
     "knip:no-cache": "knip",
     "locale": "lobe-i18n locale",
@@ -36,6 +36,9 @@
     "json-schema": "tsx scripts/generate-json-schema.ts",
     "storybook": "nx storybook -p 6006",
     "build-storybook": "storybook build"
+  },
+  "config": {
+    "eslint_concurrency": "4"
   },
   "devDependencies": {
     "@eslint/js": "^9.8.0",