## Summary

Switch ESLint concurrency configuration from `.env` to npm package config while keeping the default at `4`. This removes the TypeScript wrapper and uses `eslint` directly in scripts, sourcing the concurrency value from `package.json` under `config.eslint_concurrency`.

## What Changed

- Remove wrapper: delete `scripts/run-eslint.ts`.
- Scripts: update `lint` scripts to call `eslint` directly with `--concurrency=$npm_package_config_eslint_concurrency`.
- Defaults: add `"config": { "eslint_concurrency": "4" }` to `package.json` (matches the current default of 4).
- Docs: update `.env_example` to remove `ESLINT_CONCURRENCY` and note the npm config approach.

## Why

- Avoids a custom TS runner just to inject env; leverages npm’s built-in package `config`.
- Keeps a single, explicit source of truth for default concurrency in `package.json`.
- Maintains the same default value (4) introduced in the latest commit.

## How to Override

- One-off override: `pnpm lint -- --concurrency=6`
- Project default: `pnpm pkg set config.eslint_concurrency=6` (or `npm pkg set config.eslint_concurrency=6`)

## Notes

- `lint-staged` continues to run `eslint` directly (no change to its behavior). If we want lint-staged to use concurrency, we can add `--concurrency=$npm_package_config_eslint_concurrency` there in a follow-up.
- CI/CD: no behavioral change besides not reading `ESLINT_CONCURRENCY`; caching and flags remain the same.
- Migration: if you had `ESLINT_CONCURRENCY` in `.env`, it’s now ignored; remove it to avoid confusion.