@benceruleanlu 2025-09-09
I wouldn't dismiss the performance improvements of tuning this flag per-developer, especially given how often eslint is ran during development and precommit.

Regardless, I truly do not see any complexity being added, with the entire point of package scripts were to simplify and conventionalize actions across projects. Developers can continue with normal `pnpm run lint`/`pnpm run lint:fix` commands.

For maximum compatibility, the default has been changed to 4 due to stability concerns seen with auto.

**Normal**

Concurrency | Duration (s) | Per Core (s/core) | Speedup | % Reduction vs 1 Core | Δ vs Prev (s)
-- | -- | -- | -- | -- | --
1 | 113.023 | 113.023 | 1.00x | 0.0% | –
2 | 65.908 | 32.954 | 1.71x | 41.7% | -47.115
4 | 40.467 | 10.117 | 2.79x | 64.2% | -25.441
8 | 28.379 | 3.547 | 3.98x | 74.9% | -12.088

**Fix**

Concurrency | Duration (s) | Per Core (s/core) | Speedup | % Reduction vs 1 Core | Δ vs Prev (s)
-- | -- | -- | -- | -- | --
1 | 116.253 | 116.253 | 1.00x | 0.0% | –
2 | 66.511 | 33.255 | 1.75x | 42.8% | -49.742
4 | 40.625 | 10.156 | 2.86x | 65.1% | -25.886
8 | 28.515 | 3.564 | 4.08x | 75.5% | -12.110

<details>
<summary>Full Performance Test</summary>

```text
===== ESLint Performance Test =====
Started: 2025-09-08 21:29:32
Node: v24.6
OS: linux 6.6.87.2-microsoft-standard-WSL2 (x64)
CPUs: 32
Git: 0bcd9fdc9
Matrix (no cache): concurrency=[1, 2, 4, 8], fix=[false,true]
Command base: pnpm exec tsx scripts/run-eslint.ts


===== FIX=false =====
Running: concurrency=1, fix=false @ 2025-09-08 21:29:32 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-1.log
Result: concurrency=1, fix=false | ok=true exit=0 duration=113.023s perCore=113.023s/core speedup=1.00x pct_reduction_vs_1core=0.0% | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-1.log

Running: concurrency=2, fix=false @ 2025-09-08 21:31:27 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-2.log
Result: concurrency=2, fix=false | ok=true exit=0 duration=65.908s perCore=32.954s/core speedup=1.71x pct_reduction_vs_1core=41.7% delta_vs_prev=-47.115s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-2.log

Running: concurrency=4, fix=false @ 2025-09-08 21:32:35 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-4.log
Result: concurrency=4, fix=false | ok=true exit=0 duration=40.467s perCore=10.117s/core speedup=2.79x pct_reduction_vs_1core=64.2% delta_vs_prev=-25.441s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-4.log

Running: concurrency=8, fix=false @ 2025-09-08 21:33:16 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-8.log
Result: concurrency=8, fix=false | ok=true exit=0 duration=28.379s perCore=3.547s/core speedup=3.98x pct_reduction_vs_1core=74.9% delta_vs_prev=-12.088s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-false--concurrency-8.log


===== FIX=true =====
Running: concurrency=1, fix=true @ 2025-09-08 21:33:45 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-1.log
Result: concurrency=1, fix=true | ok=true exit=0 duration=116.253s perCore=116.253s/core speedup=1.00x pct_reduction_vs_1core=0.0% | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-1.log

Running: concurrency=2, fix=true @ 2025-09-08 21:35:44 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-2.log
Result: concurrency=2, fix=true | ok=true exit=0 duration=66.511s perCore=33.255s/core speedup=1.75x pct_reduction_vs_1core=42.8% delta_vs_prev=-49.742s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-2.log

Running: concurrency=4, fix=true @ 2025-09-08 21:36:52 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-4.log
Result: concurrency=4, fix=true | ok=true exit=0 duration=40.625s perCore=10.156s/core speedup=2.86x pct_reduction_vs_1core=65.1% delta_vs_prev=-25.886s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-4.log

Running: concurrency=8, fix=true @ 2025-09-08 21:37:33 | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-8.log
Result: concurrency=8, fix=true | ok=true exit=0 duration=28.515s perCore=3.564s/core speedup=4.08x pct_reduction_vs_1core=75.5% delta_vs_prev=-12.110s | log=runs/eslint-perf-2025-09-08_21-29-32--fix-true--concurrency-8.log

Finished: 2025-09-08 21:38:02
```
</details>

<details>
<summary>Performance Testing Script</summary>

```ts
import { spawnSync } from 'node:child_process'
import fs from 'node:fs'
import os from 'node:os'
import path from 'node:path'

type RunResult = {
  ok: boolean
  exitCode: number | null
  signal: NodeJS.Signals | null
  durationMs: number
  stdout: string
  stderr: string
  error?: string
}

const pad = (n: number) => (n < 10 ? `0${n}` : `${n}`)
const fmtDateTime = (d = new Date()) =>
  `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())} ${pad(d.getHours())}:${pad(d.getMinutes())}:${pad(d.getSeconds())}`

const hrtimeMs = () => Number(process.hrtime.bigint() / 1000000n)

const headerLine = (text: string) => `\n===== ${text} =====\n`

const write = (fd: number, text: string) => fs.writeSync(fd, text)

const ensureDir = (p: string) => {
  if (!fs.existsSync(p)) fs.mkdirSync(p, { recursive: true })
}

const detectGitSha = (): string | null => {
  try {
    const out = spawnSync('git', ['rev-parse', '--short', 'HEAD'], {
      encoding: 'utf8'
    })
    if (out.status === 0) return out.stdout.trim()
    return null
  } catch {
    return null
  }
}

const runOne = (fix: boolean, concurrency: string): RunResult => {
  const args = [
    'exec',
    'tsx',
    'scripts/run-eslint.ts',
    '--concurrency',
    concurrency
  ]
  if (fix) args.push('--fix')

  const start = hrtimeMs()
  const child = spawnSync('pnpm', args, {
    encoding: 'utf8',
    maxBuffer: 10 * 1024 * 1024
  })
  const end = hrtimeMs()

  return {
    ok: child.status === 0,
    exitCode: child.status,
    signal: child.signal,
    durationMs: end - start,
    stdout: child.stdout || '',
    stderr: child.stderr || '',
    error: child.error ? String(child.error) : undefined
  }
}

const main = () => {
  const matrix = ['1', '2', '4', '8'] as const
  const fixModes = [false, true] as const

  const outDir = path.resolve('perf')
  const runsDir = path.join(outDir, 'runs')
  ensureDir(outDir)
  ensureDir(runsDir)
  const ts = new Date()
  const tsSlug = `${ts.getFullYear()}-${pad(ts.getMonth() + 1)}-${pad(ts.getDate())}_${pad(ts.getHours())}-${pad(ts.getMinutes())}-${pad(ts.getSeconds())}`
  const outPath = path.join(outDir, `eslint-perf-${tsSlug}.log`)
  const fd = fs.openSync(outPath, 'a')

  try {
    // Print early so users know where logs will be
    console.log(`ESLint performance logs writing to: ${outPath}`)
    write(fd, headerLine('ESLint Performance Test'))
    write(fd, `Started: ${fmtDateTime()}\n`)
    write(fd, `Node: ${process.version}\n`)
    write(fd, `OS: ${os.platform()} ${os.release()} (${os.arch()})\n`)
    write(fd, `CPUs: ${os.cpus().length}\n`)
    const sha = detectGitSha()
    if (sha) write(fd, `Git: ${sha}\n`)
    write(
      fd,
      `Matrix (no cache): concurrency=[${matrix.join(', ')}], fix=[false,true]\n`
    )
    write(fd, `Command base: pnpm exec tsx scripts/run-eslint.ts\n`)
    write(fd, `\n`)

    // Track baseline (concurrency=1) and previous timing for deltas and stats
    const baselineSecs: Record<string, number> = {}
    const prevSecs: Record<string, number> = {}

    for (const fix of fixModes) {
      write(fd, headerLine(`FIX=${fix}`))
      for (const conc of matrix) {
        const label = `concurrency=${conc}, fix=${fix}`
        const runLogName = `eslint-perf-${tsSlug}--fix-${fix}--concurrency-${conc}.log`
        const runLogPath = path.join(runsDir, runLogName)

        // Write to summary before running so progress is visible even if crash occurs
        write(fd, `Running: ${label} @ ${fmtDateTime()} | log=runs/${runLogName}\n`)
        try {
          fs.fsyncSync(fd)
        } catch {}

        const runFd = fs.openSync(runLogPath, 'a')
        try {
          write(runFd, headerLine('ESLint Performance Run'))
          write(runFd, `Started: ${fmtDateTime()}\n`)
          write(runFd, `Label: ${label}\n`)
          write(runFd, `Node: ${process.version}\n`)
          write(
            runFd,
            `OS: ${os.platform()} ${os.release()} (${os.arch()}) | CPUs: ${os.cpus().length}\n`
          )
          const shaRun = detectGitSha()
          if (shaRun) write(runFd, `Git: ${shaRun}\n`)
          write(
            runFd,
            `Command: pnpm exec tsx scripts/run-eslint.ts --concurrency ${conc}${fix ? ' --fix' : ''}\n`
          )
          write(runFd, `\n`)

          // Ensure run header is flushed before executing, in case the run crashes
          try {
            fs.fsyncSync(runFd)
          } catch {}

          const res = runOne(fix, conc)

          // Compute and write stats
          const secs = res.durationMs / 1000
          const cores = Number(conc)
          const perCore = secs / (cores || 1)
          const fixKey = `fix=${fix}`
          if (conc === '1' && res.ok) baselineSecs[fixKey] = secs
          const base = baselineSecs[fixKey]
          const pctReduction = base ? ((base - secs) / base) * 100 : 0
          const speedup = base ? base / secs : 1
          const prevKey = `prev-${fixKey}`
          const prev = prevSecs[prevKey]
          const deltaPrev = prev ? secs - prev : 0
          prevSecs[prevKey] = secs

          write(
            fd,
            `Result: ${label} | ok=${res.ok} exit=${res.exitCode} duration=${secs.toFixed(3)}s perCore=${perCore.toFixed(3)}s/core` +
              (base ? ` speedup=${speedup.toFixed(2)}x pct_reduction_vs_1core=${pctReduction.toFixed(1)}%` : '') +
              (prev ? ` delta_vs_prev=${deltaPrev >= 0 ? '+' : ''}${deltaPrev.toFixed(3)}s` : '') +
              ` | log=runs/${runLogName}\n`
          )
          try {
            fs.fsyncSync(fd)
          } catch {}

          write(
            runFd,
            `Result: ok=${res.ok} exit=${res.exitCode} signal=${res.signal} duration=${secs.toFixed(3)}s perCore=${perCore.toFixed(3)}s/core` +
              (base ? ` speedup=${speedup.toFixed(2)}x pct_reduction_vs_1core=${pctReduction.toFixed(1)}%` : '') +
              (prev ? ` delta_vs_prev=${deltaPrev >= 0 ? '+' : ''}${deltaPrev.toFixed(3)}s` : '') +
              `\n`
          )
          if (res.error) write(runFd, `Error: ${res.error}\n`)
          if (res.stdout) {
            write(runFd, `--- stdout ---\n`)
            write(runFd, res.stdout)
            if (!res.stdout.endsWith('\n')) write(runFd, '\n')
          }
          if (res.stderr) {
            write(runFd, `--- stderr ---\n`)
            write(runFd, res.stderr)
            if (!res.stderr.endsWith('\n')) write(runFd, '\n')
          }
          write(runFd, `Finished: ${fmtDateTime()}\n`)
        } finally {
          try {
            fs.fsyncSync(runFd)
          } catch {}
          fs.closeSync(runFd)
        }
        write(fd, `\n`)
        try {
          fs.fsyncSync(fd)
        } catch {}
      }
    }

    write(fd, `Finished: ${fmtDateTime()}\n`)
  } finally {
    fs.closeSync(fd)
  }

  console.log(`ESLint performance log saved to: ${outPath}`)
}

main()
```
</details>